Renamed power combiner to electricity combiner
